requirements.txt
veracode-api-py>=0.9.13
logging-formatter-anticrlf>=1.2